+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/types/SOUNDENVELOPE.java
@@ -12,7 +12,8 @@
  * Lesser General Public License for more details.
  * 
  * You should have received a copy of the GNU Lesser General Public
- * License along with this library. */
+ * License along with this library.
+ */
 package com.jpexs.decompiler.flash.types;
 
 import com.jpexs.decompiler.flash.types.annotations.SWFType;
@@ -32,4 +33,38 @@ public class SOUNDENVELOPE implements Serializable {
 
     @SWFType(BasicType.UI16)
     public int rightLevel;
+
+    @Override
+    public int hashCode() {
+        int hash = 7;
+        hash = 29 * hash + (int) (this.pos44 ^ (this.pos44 >>> 32));
+        hash = 29 * hash + this.leftLevel;
+        hash = 29 * hash + this.rightLevel;
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        final SOUNDENVELOPE other = (SOUNDENVELOPE) obj;
+        if (this.pos44 != other.pos44) {
+            return false;
+        }
+        if (this.leftLevel != other.leftLevel) {
+            return false;
+        }
+        if (this.rightLevel != other.rightLevel) {
+            return false;
+        }
+        return true;
+    }
+
 }